views: put API views in its own file
